Move some CodeMirror code.
